[#issue-84] subscriber should manage heap dump and parsing thread
